replace f64 to BigDecimal

+++ src/config.rs
@@ -1,4 +1,5 @@
 use crate::error::*;
+use bigdecimal::BigDecimal;
 use serde::{Deserialize, Serialize};
 use std::fmt;
 use std::fs;
@@ -46,7 +47,7 @@ pub struct Config {
     pub debug: bool,                // デバッグ出力の有無
     pub ast_format: AstFormat,      // ASTのフォーマット
     pub trig_func_arg: TrigFuncArg, // 三角関数の引数, 逆三角関数の結果
-    pub log_base: f64,              // logの底
+    pub log_base: BigDecimal,       // logの底
     pub num_of_digit: u32,          // 結果の小数点以下の桁数
 }
